(editor) Move resetPhrases helper into phrases section module
