+++ src/editor/sections/phrases.js
@@ -4,7 +4,9 @@
 // ------------------------------------------------------------------ //
 
 import { html } from "lit";
-import { detectLang, SUPPORTED_LOCALES } from "../../i18n.js";
+import { t, detectLang, SUPPORTED_LOCALES } from "../../i18n.js";
+import { normalize } from "../../utils/normalize.js";
+import { toCanonicalAllergenKey } from "../../constants.js";
 
 export function renderPhrasesSection(editor) {
   const c = editor._editorConfig();
@@ -173,4 +175,63 @@ export function renderPhrasesSection(editor) {
       </ha-formfield>
     </details>
   `;
+}
+
+/**
+ * Seed config.phrases with the localized default names for the chosen
+ * language (the "Apply translations" action), and set date_locale to it.
+ * Shared by the card and badge editors. Uses _currentAllergens() so the
+ * GPL/GP installed plants are included via the editor's discovered list.
+ */
+export function resetPhrases(editor, lang) {
+  if (editor.debug) console.debug("[Editor] resetPhrases - lang:", lang);
+  editor._updateConfig("date_locale", lang);
+
+  const integration = editor._config?.integration;
+  const rawKeys = editor._currentAllergens();
+
+  const full = {};
+  const short = {};
+  rawKeys.forEach((raw) => {
+    const normKey = normalize(raw);
+    const canonKey = toCanonicalAllergenKey(normKey);
+    // SILAM's aggregate uses the user-facing 'index' name, not 'allergy_risk'.
+    const transKey = normKey === "index" ? "index" : canonKey;
+    // Use the shared resolver so a missing editor phrase falls back to the
+    // card allergen name / capitalized term instead of storing the raw i18n
+    // key as the phrase default (issue #262 follow-up: `graminales`).
+    full[raw] = editor._resolveAllergenPhrase(transKey, raw, { lang });
+    short[raw] = editor._resolveAllergenPhrase(transKey, raw, {
+      short: true,
+      lang,
+    });
+  });
+
+  const numLevels = editor._currentNumLevels();
+  // 5-level integrations (MSW, PEU, Kleenex) use the scale-specific severity
+  // labels rather than the first five of the 7-level palette.
+  const levelKeyPrefix =
+    integration === "msw" ||
+    integration === "irmkmi" ||
+    integration === "peu" ||
+    integration === "kleenex"
+      ? "editor.phrases_levels5"
+      : "editor.phrases_levels";
+  const levels = Array.from({ length: numLevels }, (_, i) =>
+    t(`${levelKeyPrefix}.${i}`, lang),
+  );
+
+  const days = {
+    0: t(`editor.phrases_days.0`, lang),
+    1: t(`editor.phrases_days.1`, lang),
+    2: t(`editor.phrases_days.2`, lang),
+  };
+
+  editor._updateConfig("phrases", {
+    full,
+    short,
+    levels,
+    days,
+    no_information: t("editor.no_information", lang),
+  });
 }
